Finished README and testcases, tar'd PA3 directory for submission.

+++ PA3/PA2Test2.java
@@ -4,7 +4,7 @@ import meggy.Meggy;
 {
 	public static void main(String[] test)
 	{
-		Meggy.setPixel(1, 1, Meggy.Color.RED);
+		Meggy.setPixel((byte)1, (byte)1, Meggy.Color.RED);
 	}
 }
